requirements.txt
requests>=2.28.0
dnspython>=2.4.0
boto3>=1.28.0
